refactor(mcp): extract monolith index.ts into modular handlers
Split the 2,348-line MCP server monolith into focused modules:
- types.ts: shared interfaces (FirewatchParams, McpToolResult)
- handlers/: query, status, feedback, mutations, config, help
- context/: repo, mutation, feedback context factories
- utils/: formatting, parsing, sync, id-resolution
- index.ts slimmed to ~290 lines (server class + tool registration)

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
